requirements.txt
flask-sqlalchemy>=3.1.1
flask>=3.0.3
psycopg2-binary>=2.9.9
werkzeug>=3.0.4